back-end/requirements.txt
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-extensions==3.2.3
asgiref==3.7.2
python-dotenv==1.0.0
psycopg2==2.9.6
pillow==10.0.0
django-cors-headers==4.2.0
